+++ test/int_types_test.ml
@@ -10,7 +10,7 @@ let test_signed (type t) ~(create : int -> t) (module T : Test_lib.T with type t
 
 let test_unsigned (type t) ~(create : int -> t) (module T : Test_lib.T with type t = t) =
   Printf.printf "Test %s\n%!" (T.name' ());
-  let values = [0; 1; 2; 2147483647] in
+  let values = [0; 1; 2; 2147483647; 4294967295] in
   List.iter
     ~f:(fun v -> Test_lib.test_encode (module T) (create v))
     values